MDL-74259 theme_boost: swap order of form save/cancel buttons.
The primary action button(s) should be to the right of the cancel
button. Achieve this by ensuring cancel is always ordered first
regardless of order within DOM.

By leaving the DOM ordering intact, we ensure the first defined
action button is still the one regarded as default for the form.

Co-authored-by: Mikel Martín <mikel@moodle.com>